% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_tail_weights}
\alias{calc_tail_weights}
\title{Calculate tail weights}
\usage{
calc_tail_weights(x, knots)
}
\arguments{
\item{x}{grid of points at which to evaluate weights}

\item{knots}{thresholds above which GPDs are defined}
}
\value{
weight matrix for each GPD model along grid \code{x}
}
\description{
Calculate tail weights for weighted GPDs
}
\examples{
xs <- seq(0, 10, l = 10000)
w = calc_tail_weights(xs, knots= 1:5)
matplot(xs,w, type = "l")
}
