% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_weight_over_x}
\alias{calc_weight_over_x}
\title{Calculate triangular weights over a grid}
\usage{
calc_weight_over_x(x, st, md, ed)
}
\arguments{
\item{x}{grid of points on which to evaluate weights}

\item{st}{starting point (see description)}

\item{md}{midpoint (see description)}

\item{ed}{end point (see description)}
}
\value{
estimates of triangular weights at \code{x}
}
\description{
Calculate triangular weights over a grid
}
