% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_weighted_quantile}
\alias{calc_weighted_quantile}
\title{Calculate weighted quantile}
\usage{
calc_weighted_quantile(qs, taus, thresh.taus, end.tau, type = c("upper",
  "lower"))
}
\arguments{
\item{qs}{quantile estimates from different QR and GPD models}

\item{taus}{quantiles to estimate}

\item{thresh.taus}{vector of quantiles, above which the GPDs are defined}

\item{end.tau}{Last quantile, above which the outer most GPD gets all 
the weight and the remaining have weight zero}

\item{type}{One of "upper" or "lower" depending on if upper or lower 
quantiles are being estimated}
}
\value{
weighted estimate of a quantile
}
\description{
Estimate a weighted quantile given several estimates of the
same quantile from different QR and GPD models
}
