% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_mixture.R
\name{fgpdset}
\alias{fgpdset}
\title{Fit collection of GPD models to compounding exceedances as defined in 
\code{get_exceedances}}
\usage{
fgpdset(excd.list, Xb, type = c("upper", "lower"), return.qs = FALSE,
  pred.taus = NULL, use.winter.shape = FALSE, all.winter.ind = NULL,
  Xb.pred.time = NULL)
}
\arguments{
\item{excd.list}{list of exceedances created by \code{get_exceedances}}

\item{Xb}{basis design matrix for tails (log(scale) of GPD). Each basis 
function is in a  separate column.}

\item{type}{one of "upper" or "lower" if fitting upper or lower tails}

\item{return.qs}{(logical) should quantile estimates be returned}

\item{pred.taus}{vector of quantiles to predict}

\item{use.winter.shape}{(logical) should a separate shape parameter be 
estimated for the winter months?}

\item{all.winter.ind}{indicator for which dates correspond to winter months}

\item{Xb.pred.time}{Basis matrix for prediction dates.}
}
\value{
List of fitted parameter matrix (one column for each GPD order: log(scale), 
shape) and estimated quantiles (qs, of dimension time, taus, model).
}
\description{
Fit collection of GPD models to compounding exceedances as defined in 
\code{get_exceedances}
}
