% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_mixture.R
\name{fit_weighted_rqgpd}
\alias{fit_weighted_rqgpd}
\title{Fit a weighted quantile regression and gpd model}
\usage{
fit_weighted_rqgpd(y, dates, b.df.lf, b.df.hf, b.b.formula, t.df.lf, t.df.hf,
  t.b.formula, taus, l.taus, u.taus, thresh.l.taus, thresh.u.taus,
  use.winter.shape = FALSE, pred.dates = NULL)
}
\arguments{
\item{y}{response vector samples}

\item{dates}{dates corresponding to response observations}

\item{b.df.lf}{df for bulk low freq basis}

\item{b.df.hf}{df for bulk high freq basis}

\item{b.b.formula}{bulk formula involving b.lf and b.hf}

\item{t.df.lf}{df for tail low freq basis}

\item{t.df.hf}{df for tail high freq basis}

\item{t.b.formula}{tail formula (simpler) involving b.lf and b.hf}

\item{taus}{vector of intermediate (bulk) quantiles to estimate 
(between 0 and 1)}

\item{l.taus}{vector of lower tail quantiles to estimate (between 0 and 1)}

\item{u.taus}{vector of higher tail quantiles to estimate (between 0 and 1)}

\item{thresh.l.taus}{lower threshold quantiles for defining collection of 
GPDs}

\item{thresh.u.taus}{upper threshold quantiles for defining collection of 
GPDs}

\item{use.winter.shape}{(logical) should a separate shape parameter be 
estimated for winter months?}

\item{pred.dates}{(optional) vector of dates to make predictions on 
(if different from dates)}
}
\description{
Estimate quantiles using mixture of quantile regression and 
GPDs. Use triangular weighting functions to tradeoff between QR fit and GPD
fits.
}
