% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_mixture.R
\name{get_exceedances}
\alias{get_exceedances}
\title{Extract compounding threshold exceedances above outer QR quantiles and 
GPD quantiles}
\usage{
get_exceedances(y, dates, df.lf, df.hf, high.taus = NULL, low.taus = NULL,
  bulk.Xb = NULL, tail.Xb = NULL, tail.df.lf = df.lf,
  tail.df.hf = df.hf, bulk.formula = ~b.lf + b.hf + b.lf:b.hf,
  tail.formula = ~b.lf + b.hf + b.lf:b.hf, use.winter.shape = FALSE)
}
\arguments{
\item{y}{response vector}

\item{dates}{vector of dates of each observation}

\item{df.lf}{natural spline degrees of freedom on date}

\item{df.hf}{periodic basis spline degrees of freedom on the 
day of the year}

\item{high.taus}{Upper quantile thresholds for fitting GPDs and extracting 
exceedances}

\item{low.taus}{Lower quantile thresholds for fitting GPDs and extracting 
exceedances}

\item{bulk.Xb}{Basis function matrix for bulk of distribution}

\item{tail.Xb}{Basis function matrix for tails of distribution}

\item{tail.df.lf}{natural spline degrees of freedom on date for tail}

\item{tail.df.hf}{periodic basis spline degrees of freedom on the 
day of the year for tail}

\item{bulk.formula}{R formula for quantile regression on bulk}

\item{tail.formula}{R formula for log(scale) GPD parameters in tails}

\item{use.winter.shape}{(logical) Should a separate shape parameter be fit
for the winter and rest?}
}
\value{
List containing exceedances above and below the high.taus and 
low.taus quantiles of the temperature distributions as estimated by QR and 
GPD models
}
\description{
Extract compounding threshold exceedances above outer QR quantiles and 
GPD quantiles
}
