% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_mixture.R
\name{get_gp_pars}
\alias{get_gp_pars}
\title{Get GPD parameters}
\usage{
get_gp_pars(bpars, Xb, winter.ind = NULL)
}
\arguments{
\item{bpars}{basis coefficients in same order as columns of Xb (basis matrix)}

\item{Xb}{basis function design matrix for prediction time points}

\item{winter.ind}{if a separater shape should be used for the winter months,
this is an indicator for which time points correspond to winter months.}
}
\value{
list of fitted GPD parameters at each time point (scale and shape)
}
\description{
Get GPD parameters from matrix of basis coefficients and basis function
design matrix.
}
