% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_basis.R
\name{make_time_basis}
\alias{make_time_basis}
\title{Make time basis matrix}
\usage{
make_time_basis(dates, df.lf, df.hf, b.formula = ~b.lf + b.hf + b.lf:b.hf)
}
\arguments{
\item{dates}{vector of dates of each observation}

\item{df.lf}{natural spline degrees of freedom on date}

\item{df.hf}{periodic basis spline degrees of freedom on the 
day of the year}

\item{b.formula}{formula for natural spline basis and periodic basis 
splines. Must be written in terms of b.lf and b.hf}
}
\value{
matrix of basis functions with all linear and interaction terms 
        evaluated at every date passed to function.
}
\description{
makes a matrix of polynomial (in date) and natural spline
(in day of year) basis functions + interaction terms for set of dates.
Dates are scaled to be between (0,1) before defining splines.
}
