% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_winter_ind}
\alias{make_winter_ind}
\title{Make winter indicator}
\usage{
make_winter_ind(dates)
}
\arguments{
\item{dates}{vector of dates to create indicator for}
}
\value{
indicator of which dates are in winter
}
\description{
Create an indicator for whether a given date is between 
December and February
}
