% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_mixture.R
\name{qtailgpd}
\alias{qtailgpd}
\title{Calculate tail quantiles from GPD}
\usage{
qtailgpd(taus, scale, shape, threshold = NULL, thresh_tau = NULL,
  type = c("upper", "lower"))
}
\arguments{
\item{taus}{vector of quantiles to estimate (refers to cumulative probs of
entire distribution, not just distribution of exceedances)}

\item{scale}{Vector of fitted scale parameters, one for each date.}

\item{shape}{Vector of fitted shape parameters, one for each date.}

\item{threshold}{Vector of thresholds above which the GPD was estimated}

\item{thresh_tau}{Corresponding estimate of CDF at threshold parameters}

\item{type}{(string) either "upper" for upper tail or "lower" for lower tail.}
}
\value{
matrix of quantile estimates (rows: dates, columns: taus)
}
\description{
Calculate tail quantiles from GPD
}
