% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_regression.R
\name{rqbulk}
\alias{rqbulk}
\title{Quantile regression for bulk of distribution}
\usage{
rqbulk(y, dates, df.lf, df.hf, taus, Xb = NULL, b.formula = ~b.lf + b.hf +
  b.lf:b.hf, threshold.tau = NULL, return.basis = TRUE,
  return.yhat = TRUE, return.excd = FALSE, type = c("upper", "lower"))
}
\arguments{
\item{y}{response vector}

\item{dates}{vector of dates of each observation}

\item{df.lf}{natural spline degrees of freedom on date}

\item{df.hf}{periodic basis spline degrees of freedom on the 
day of the year}

\item{taus}{quantiles to estimate}

\item{Xb}{(optional) matrix of basis functions. Each basis function is in a 
separate column.}

\item{b.formula}{basis formula. Should be function of b.lf and b.hf}

\item{threshold.tau}{quantile above or below which taus are to be interpreted 
conditionally for distributions exceeding threshold.tau (either below or
above).}

\item{return.basis}{(logical) should the basis matrix be returned?}

\item{return.yhat}{(logical) should matrix of predicted quantiles (see taus)
be returned?}

\item{return.excd}{(logical) should vector of exceedances of highest quantile
be returned?}

\item{type}{one of "upper" or "lower" if tail tau quantiles should be handled
separately.}
}
\value{
A list containing the basis matrix (Xb), fitted quantiles (yhat), QR 
basis coefficients (betas), lower and upper exceedances, dates, indices, and 
thresholds (lexcd, lexcd.dates, lexcd.id, lthresh) and (uexcd, uexcd.dates, 
uexcd.id, uthresh)
}
\description{
Fit quantile regression to bulk of data and extract exceedances
}
