% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_regression.R
\name{rqcomb}
\alias{rqcomb}
\title{Combined quantile regression}
\usage{
rqcomb(y, dates, b.df.lf, b.df.hf, b.b.formula, t.df.lf, t.df.hf, t.b.formula,
  taus, l.taus, u.taus, pred.dates = NULL)
}
\arguments{
\item{y}{response vector}

\item{dates}{vector of dates of each observation}

\item{b.df.lf}{df for bulk low freq basis}

\item{b.df.hf}{df for bulk high freq basis}

\item{b.b.formula}{bulk formula involving b.lf and b.hf}

\item{t.df.lf}{df for tail low freq basis}

\item{t.df.hf}{df for tail high freq basis}

\item{t.b.formula}{tail formula (simpler) involving b.lf and b.hf}

\item{taus}{vector of intermediate (bulk) quantiles to estimate 
(between 0 and 1)}

\item{l.taus}{vector of lower tail quantiles to estimate (between 0 and 1)}

\item{u.taus}{vector of higher tail quantiles to estimate (between 0 and 1)}

\item{pred.dates}{(optional) vector of dates to make predictions on 
(if different from dates)}
}
\value{
List of lower, mid, and upper quantile estimates
}
\description{
Fit different quantile regression models to the bulk quantiles
and the tail quantiles. This allows for a higher order basis design in
the bulk quantiles and a lower order basis design in the tails.
}
