% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{weight_one}
\alias{weight_one}
\title{Weight for single observation}
\usage{
weight_one(x, st, md, ed)
}
\arguments{
\item{x}{input point to evaluate weights at}

\item{st}{starting point (see description)}

\item{md}{midpoint (see description)}

\item{ed}{end point (see description)}
}
\description{
Calculate weights for triangular weight functions, which 
increase from 0 to 1 between st and md, and decrease from 1 to 0 from md to
ed.
}
